refactor: remove duplication of functions that get csrf token and cookies
